Cognitive Routing & RAG/requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-core>=0.2.0
langgraph>=0.1.19
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
numpy>=1.26.0
